+++ src/main/java/webserver/http/BodyStatus.java
@@ -1,4 +1,4 @@
-package utils;
+package webserver.http;
 
 public enum BodyStatus {
     EXISTS(true),
