+++ src/main/java/webserver/http/request/RequestBody.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.request;
 
 import utils.QueryParser;
